+++ docs-site/versioned_docs/version-3.4.0/api/typedoc-sidebar.cjs
@@ -1,0 +1,328 @@
+// @ts-check
+/** @type {import("@docusaurus/plugin-content-docs").SidebarsConfig} */
+const typedocSidebar = {
+  items: [
+    {
+      type: "category",
+      label: "Enumerations",
+      items: [
+        {
+          type: "doc",
+          id: "api/enumerations/DriverEnum",
+          label: "DriverEnum"
+        },
+        {
+          type: "doc",
+          id: "api/enumerations/FilterOperatorEnum",
+          label: "FilterOperatorEnum"
+        },
+        {
+          type: "doc",
+          id: "api/enumerations/PaginationModeEnum",
+          label: "PaginationModeEnum"
+        },
+        {
+          type: "doc",
+          id: "api/enumerations/SortEnum",
+          label: "SortEnum"
+        }
+      ]
+    },
+    {
+      type: "category",
+      label: "Classes",
+      items: [
+        {
+          type: "doc",
+          id: "api/classes/InvalidFilterOperatorValueError",
+          label: "InvalidFilterOperatorValueError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/InvalidLimitError",
+          label: "InvalidLimitError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/InvalidPageNumberError",
+          label: "InvalidPageNumberError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/InvalidResourceNameError",
+          label: "InvalidResourceNameError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/JsonApiRequestStrategy",
+          label: "JsonApiRequestStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/JsonApiResponseStrategy",
+          label: "JsonApiResponseStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/KeyNotFoundError",
+          label: "KeyNotFoundError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/LaravelRequestStrategy",
+          label: "LaravelRequestStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/LaravelResponseStrategy",
+          label: "LaravelResponseStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/NestjsRequestStrategy",
+          label: "NestjsRequestStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/NestjsResponseStrategy",
+          label: "NestjsResponseStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/NgQubeeModule",
+          label: "NgQubeeModule"
+        },
+        {
+          type: "doc",
+          id: "api/classes/NgQubeeService",
+          label: "NgQubeeService"
+        },
+        {
+          type: "doc",
+          id: "api/classes/PaginatedCollection",
+          label: "PaginatedCollection"
+        },
+        {
+          type: "doc",
+          id: "api/classes/PaginationNotSyncedError",
+          label: "PaginationNotSyncedError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/PaginationService",
+          label: "PaginationService"
+        },
+        {
+          type: "doc",
+          id: "api/classes/PostgrestRequestStrategy",
+          label: "PostgrestRequestStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/PostgrestResponseStrategy",
+          label: "PostgrestResponseStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/SpatieRequestStrategy",
+          label: "SpatieRequestStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/SpatieResponseStrategy",
+          label: "SpatieResponseStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/StrapiRequestStrategy",
+          label: "StrapiRequestStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/StrapiResponseStrategy",
+          label: "StrapiResponseStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnselectableModelError",
+          label: "UnselectableModelError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnsupportedFieldSelectionError",
+          label: "UnsupportedFieldSelectionError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnsupportedFilterError",
+          label: "UnsupportedFilterError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnsupportedFilterOperatorError",
+          label: "UnsupportedFilterOperatorError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnsupportedIncludesError",
+          label: "UnsupportedIncludesError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnsupportedSearchError",
+          label: "UnsupportedSearchError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnsupportedSelectError",
+          label: "UnsupportedSelectError"
+        },
+        {
+          type: "doc",
+          id: "api/classes/UnsupportedSortError",
+          label: "UnsupportedSortError"
+        }
+      ]
+    },
+    {
+      type: "category",
+      label: "Interfaces",
+      items: [
+        {
+          type: "doc",
+          id: "api/interfaces/IConfig",
+          label: "IConfig"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IFields",
+          label: "IFields"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IFilters",
+          label: "IFilters"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/INestState",
+          label: "INestState"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IOperatorFilter",
+          label: "IOperatorFilter"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IPage",
+          label: "IPage"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IPaginatedObject",
+          label: "IPaginatedObject"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IPaginationConfig",
+          label: "IPaginationConfig"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IQueryBuilderConfig",
+          label: "IQueryBuilderConfig"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IQueryBuilderState",
+          label: "IQueryBuilderState"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IRequestStrategy",
+          label: "IRequestStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/IResponseStrategy",
+          label: "IResponseStrategy"
+        },
+        {
+          type: "doc",
+          id: "api/interfaces/ISort",
+          label: "ISort"
+        }
+      ]
+    },
+    {
+      type: "category",
+      label: "Type Aliases",
+      items: [
+        {
+          type: "doc",
+          id: "api/type-aliases/HeaderBag",
+          label: "HeaderBag"
+        }
+      ]
+    },
+    {
+      type: "category",
+      label: "Variables",
+      items: [
+        {
+          type: "doc",
+          id: "api/variables/NG_QUBEE_DRIVER",
+          label: "NG_QUBEE_DRIVER"
+        },
+        {
+          type: "doc",
+          id: "api/variables/NG_QUBEE_REQUEST_OPTIONS",
+          label: "NG_QUBEE_REQUEST_OPTIONS"
+        },
+        {
+          type: "doc",
+          id: "api/variables/NG_QUBEE_REQUEST_STRATEGY",
+          label: "NG_QUBEE_REQUEST_STRATEGY"
+        },
+        {
+          type: "doc",
+          id: "api/variables/NG_QUBEE_RESPONSE_OPTIONS",
+          label: "NG_QUBEE_RESPONSE_OPTIONS"
+        },
+        {
+          type: "doc",
+          id: "api/variables/NG_QUBEE_RESPONSE_STRATEGY",
+          label: "NG_QUBEE_RESPONSE_STRATEGY"
+        }
+      ]
+    },
+    {
+      type: "category",
+      label: "Functions",
+      items: [
+        {
+          type: "doc",
+          id: "api/functions/buildNgQubeeProviders",
+          label: "buildNgQubeeProviders"
+        },
+        {
+          type: "doc",
+          id: "api/functions/provideNgQubee",
+          label: "provideNgQubee"
+        },
+        {
+          type: "doc",
+          id: "api/functions/provideNgQubeeInstance",
+          label: "provideNgQubeeInstance"
+        },
+        {
+          type: "doc",
+          id: "api/functions/readHeader",
+          label: "readHeader"
+        }
+      ]
+    }
+  ]
+};
+module.exports = typedocSidebar.items;